fixed tracklist special character bug
after the artist name a dash before the song title in the track list showed two missing character boxes. Tested a bunch of UTF-8 settings to see what caused the non-printible error. While testing, one of the existing albums in the libry stripped the whole first part with the artist name and track number. Realized that was actually the better option. Dont need the artist name in the track title anyway.