switched navigation icon from feeds to import
The app was originally going to use rss to get albums and info. But the audio needed to come from m3u links and it now uses relative paths to fetch the rss info for the artist. So the nav having a feeds icon could be cofusing. Changed to import and now the add link field indicates it needs an m3u link